Reformate, Cleanup, Import organization

+++ src/main/java/info/u_team/draw_bridge/tileentity/render/DrawBridgeTileEntityRender.java
@@ -8,11 +8,7 @@ import info.u_team.draw_bridge.block.DrawBridgeBlock;
 import info.u_team.draw_bridge.tileentity.DrawBridgeTileEntity;
 import net.minecraft.block.BlockState;
 import net.minecraft.client.Minecraft;
-import net.minecraft.client.renderer.BlockModelRenderer;
-import net.minecraft.client.renderer.BlockRendererDispatcher;
-import net.minecraft.client.renderer.BufferBuilder;
-import net.minecraft.client.renderer.RenderHelper;
-import net.minecraft.client.renderer.Tessellator;
+import net.minecraft.client.renderer.*;
 import net.minecraft.client.renderer.texture.AtlasTexture;
 import net.minecraft.client.renderer.tileentity.TileEntityRenderer;
 import net.minecraft.client.renderer.vertex.DefaultVertexFormats;
@@ -21,12 +17,11 @@ import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
 
 public class DrawBridgeTileEntityRender extends TileEntityRenderer<DrawBridgeTileEntity> {
-
+	
 	private BlockRendererDispatcher blockRenderer = Minecraft.getInstance().getBlockRendererDispatcher();
-
+	
 	@Override
-	public void render(DrawBridgeTileEntity tileEntityIn, double x, double y, double z, float partialTicks,
-			int destroyStage) {
+	public void render(DrawBridgeTileEntity tileEntityIn, double x, double y, double z, float partialTicks, int destroyStage) {
 		BlockPos pos = tileEntityIn.getPos();
 		World world = this.getWorld();
 		Direction facing = world.getBlockState(pos).get(DrawBridgeBlock.FACING);
@@ -35,8 +30,7 @@ public class DrawBridgeTileEntityRender extends TileEntityRenderer<DrawBridgeTil
 			BufferBuilder bufferbuilder = tessellator.getBuffer();
 			this.bindTexture(AtlasTexture.LOCATION_BLOCKS_TEXTURE);
 			RenderHelper.disableStandardItemLighting();
-			GlStateManager.blendFunc(GlStateManager.SourceFactor.SRC_ALPHA,
-					GlStateManager.DestFactor.ONE_MINUS_SRC_ALPHA);
+			GlStateManager.blendFunc(GlStateManager.SourceFactor.SRC_ALPHA, GlStateManager.DestFactor.ONE_MINUS_SRC_ALPHA);
 			GlStateManager.enableBlend();
 			GlStateManager.disableCull();
 			if (Minecraft.isAmbientOcclusionEnabled()) {
@@ -44,21 +38,18 @@ public class DrawBridgeTileEntityRender extends TileEntityRenderer<DrawBridgeTil
 			} else {
 				GlStateManager.shadeModel(7424);
 			}
-
+			
 			BlockModelRenderer.enableCache();
 			bufferbuilder.begin(7, DefaultVertexFormats.BLOCK);
-
+			
 			BlockState[] states = tileEntityIn.getBlocksToRender();
 			for (int i = 0; i < states.length; i++) {
 				int off = i + 1;
 				BlockPos tmppos = pos.offset(facing, off);
-				bufferbuilder.setTranslation(
-						x - (double) tmppos.getX() + (facing.getXOffset() * off) - tileEntityIn.getOffsetX(partialTicks),
-						y - (double) tmppos.getY() + (facing.getYOffset() * off) - tileEntityIn.getOffsetY(partialTicks),
-						z - (double) tmppos.getZ() + (facing.getZOffset() * off) - tileEntityIn.getOffsetZ(partialTicks));
+				bufferbuilder.setTranslation(x - tmppos.getX() + (facing.getXOffset() * off) - tileEntityIn.getOffsetX(partialTicks), y - tmppos.getY() + (facing.getYOffset() * off) - tileEntityIn.getOffsetY(partialTicks), z - tmppos.getZ() + (facing.getZOffset() * off) - tileEntityIn.getOffsetZ(partialTicks));
 				this.renderStateModel(tmppos, states[i], bufferbuilder, world, false);
 			}
-
+			
 			bufferbuilder.setTranslation(0.0D, 0.0D, 0.0D);
 			tessellator.draw();
 			BlockModelRenderer.disableCache();
@@ -70,14 +61,12 @@ public class DrawBridgeTileEntityRender extends TileEntityRenderer<DrawBridgeTil
 	public boolean isGlobalRenderer(DrawBridgeTileEntity te) {
 		return true;
 	}
-
+	
 	@SuppressWarnings("deprecation")
-	private boolean renderStateModel(BlockPos pos, BlockState state, BufferBuilder buffer, World world,
-			boolean checkSides) {
+	private boolean renderStateModel(BlockPos pos, BlockState state, BufferBuilder buffer, World world, boolean checkSides) {
 		if (blockRenderer == null)
 			blockRenderer = Minecraft.getInstance().getBlockRendererDispatcher();
-		return this.blockRenderer.getBlockModelRenderer().renderModel(world, this.blockRenderer.getModelForState(state),
-				state, pos, buffer, checkSides, new Random(), state.getPositionRandom(pos));
+		return this.blockRenderer.getBlockModelRenderer().renderModel(world, this.blockRenderer.getModelForState(state), state, pos, buffer, checkSides, new Random(), state.getPositionRandom(pos));
 	}
-
+	
 }
